simplified build of DOM structure
Replaced the creation of separate attribute objects by the much simpler
setAttribute(), as suggested by gandro. That should be available everywhere
all the other DOM functions are, including IE 5.5+ (IE is not satisfied with
the type attribute however and is still missing the classid for the ActiveX
component, I'll fix this later).

Also abolished the bumbling use of toString(). :p