reduced flash plugin parameters; experimental IE support
Reduced number of attributes and PARAMs for the Flash object to the bare
minimum ("type", "width" and "height" attributes and "movie" parameter which is
the preferable standard to "src"). The "data" attribute is still needed for
proper support in NPAPI browsers, though.

If you want to add other parameters, please refer to Adobe's documentation:

      http://kb2.adobe.com/cps/127/tn_12701.html

Also replaced all occurrences of the Mozilla-specific const keyword with
var, allowing the script to run on IE and several older non-Mozilla Browsers
(Opera < 9 etc). Flash Player still has issues with loading films on IE,
however.